Restructure adapter-knex package
