+++ .changeset/afdfb72d/changes.json
@@ -1,0 +1,1 @@
+{ "releases": [{ "name": "@keystone-alpha/adapter-knex", "type": "patch" }], "dependents": [] }
