+++ .changeset/afdfb72d/changes.md
@@ -1,0 +1,1 @@
+- Restructure internal code
